refactor(team): extract HTTP middleware into broker_middleware.go
Slice 4 of broker.go decomposition. Moves rateLimitMiddleware,
isLivenessPath, isAgentBucketExemptPath, writeRateLimitedResponse,
rateLimitNow, consumeRateLimit, consumeAgentRateLimit,
externalWorkflowRetryAfter, pruneRateLimitEntries, rateLimitKey,
clientIPFromRequest, firstForwardedIP, trustForwardedClientIP,
setProxyClientIPHeaders, requestAuthToken, requestHasBrokerAuth,
corsMiddleware, isLoopbackRemote, hostHeaderIsLoopback,
webUIRebindGuard into a new broker_middleware.go.

requireAuth stays in broker.go because every sibling file calls
b.requireAuth(...) — moving it would multiply blast radius for no
gain. Relocates 12 middleware tests from broker_test.go into
broker_middleware_test.go and adds three new characterization tests:
TestPruneRateLimitEntries_RemovesAllAtCutoff,
TestExternalWorkflowRetryAfter_ParsesRFC3339FromErrorString,
TestRequestHasBrokerAuth_AcceptsBearerAndQueryToken.

No behavior change. Total package coverage holds at 63.0%.
